requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bootstrap4==0.1.0
boto3==1.18.11
botocore==1.21.11
Django==3.2.5
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==3.0.1
django-crispy-forms==1.12.0
jmespath==0.10.0
phonenumbers==8.12.28
pycountry==20.7.3
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
